Time: 260 ms (28.57%) | Memory: 128.2 MB (18.95%) - LeetSync

+++ 220-contains-duplicate-iii/contains-duplicate-iii.cpp
@@ -1,12 +1,11 @@
 class Solution {
 public:
     bool containsNearbyAlmostDuplicate(vector<int>& nums, int k, int z) {
-        set<long long> window;
+        set<int> window;
         for (int i = 0; i < nums.size(); i++) {
-            if (i > k) window.erase(nums[i-k-1]); // Remove numbers out of range k
-            // -z <= x - nums[i] <= z; so nums[i] - z <= x <= nums[i] + z;
-            auto it = window.lower_bound((long long) nums[i] - z);
-            if (it != window.end() && *it <= ((long long) nums[i] + z))
+            if (i > k) window.erase(nums[i-k-1]);
+            auto it = window.lower_bound( nums[i] - z);
+            if (it != window.end() && *it <= ( nums[i] + z))
                 return true;
             window.insert(nums[i]);
         }
